Introduce automatic dependency checking.
We use ./sac2c -M to get a list of dependencies for the given sac file.
Note that at the moment of writing this functionality is available for
the compiler that is built from the sources of the branch
`makedeps-follow'.

For every sac file we analyse its dependencies.  For external
dependencies (libraries outside of the current project) we
their availability at configure time.  Internal
dependencies (within the sources of the given package) are used
when producing a target for building a module for the given sac file.